move asset id generation into AssetFilter class and fix cyclic dep.
