+++ source/conf/cmake/crownstone.src.cmake
@@ -82,6 +82,8 @@ LIST(APPEND FOLDER_SOURCE "${SOURCE_DIR}/time/cs_SystemTime.cpp")
 LIST(APPEND FOLDER_SOURCE "${SOURCE_DIR}/time/cs_TimeOfDay.cpp")
 LIST(APPEND FOLDER_SOURCE "${SOURCE_DIR}/tracking/cs_TrackedDevice.cpp")
 LIST(APPEND FOLDER_SOURCE "${SOURCE_DIR}/tracking/cs_TrackedDevices.cpp")
+LIST(APPEND FOLDER_SOURCE "${SOURCE_DIR}/util/cs_AssetFilter.cpp")
+LIST(APPEND FOLDER_SOURCE "${SOURCE_DIR}/util/cs_FilterInterface.cpp")
 LIST(APPEND FOLDER_SOURCE "${SOURCE_DIR}/util/cs_ExactMatchFilter.cpp")
 LIST(APPEND FOLDER_SOURCE "${SOURCE_DIR}/util/cs_CuckooFilter.cpp")
 LIST(APPEND FOLDER_SOURCE "${SOURCE_DIR}/uart/cs_UartCommandHandler.cpp")
