+++ source/include/localisation/cs_AssetFilterPacketAccessors.h
@@ -176,28 +176,3 @@ public:
 	size_t length();
 };
 
-/**
- * Class that contains all data required for an asset filter:
- * - Runtime data.
- * - persisted data (filterdata).
- */
-class AssetFilter {
-public:
-	uint8_t* _data;  // byte representation of this object.
-	AssetFilter(uint8_t* data) : _data(data) {}
-
-	/**
-	 * Get the runtime data.
-	 */
-	asset_filter_runtime_data_t* runtimedata();
-
-	/**
-	 * Get the filter data.
-	 */
-	AssetFilterData filterdata();
-
-	/**
-	 * Get the expected length of this class.
-	 */
-	size_t length();
-};
