+++ source/include/localisation/cs_AssetFiltering.h
@@ -96,19 +96,14 @@ private:
 	 */
 	void processFilter(AssetFilter f, const scanned_device_t& asset);
 
-	/**
-	 * Returns true if the device passes the filter according to its
-	 * metadata settings. Returns false otherwise.
-	 */
-	bool filterAcceptsScannedDevice(AssetFilter filter, const scanned_device_t& asset);
 
 	/**
-	 * Returns a short_asset_id_t based on the configured selection of data
+	 * Returns a asset_id_t based on the configured selection of data
 	 * in metadata.outputType.inFormat. If the data is not sufficient, a default
 	 * constructed object is returned. (Data not sufficient can be detected:
 	 * filterInputResult will return false in that case.)
 	 */
-	asset_id_t filterOutputResultAssetId(AssetFilter filter, const scanned_device_t& asset);
+	asset_id_t getAssetId(AssetFilter filter, const scanned_device_t& asset);
 
 
 	// --------- Processing of accepted Assest ---------------
