+++ source/include/util/cs_FilterInterface.h
@@ -7,6 +7,7 @@
 
 #pragma once
 
+#include <localisation/cs_AssetFilterPacketAccessors.h>
 #include <protocol/cs_AssetFilterPackets.h>
 #include <util/cs_Crc32.h>
 
@@ -24,17 +25,10 @@ public:
 	 * In most cases a assetId is generated as crc32 from filtered input data.
 	 * This can be overwritten for special use cases like ExactMatchFilter.
 	 */
-	virtual asset_id_t assetId(const void* key, size_t keyLengthInBytes) {
-		auto crc = crc32(static_cast<const uint8_t*>(key), keyLengthInBytes, nullptr);
-
-		// little endian byte by byte copy.
-		asset_id_t id{};
-		memcpy(id.data, reinterpret_cast<uint8_t*>(&crc), 3);
-
-		return id;
-	}
-
+	virtual asset_id_t assetId(const void* key, size_t keyLengthInBytes);
 	virtual size_t size() = 0;
 
 	virtual bool isValid() = 0;
+
+
 };
